all parser functions to utils.parser
